Health block encapsulation Fixes #486
